refactor: cli 서비스, api  서비스 분리
