+++ ai/Dockerfile
@@ -33,4 +33,4 @@ ENV PYTHONPATH=/app
 ENV PYTHONUNBUFFERED=1
 
 # FastAPI 애플리케이션 실행
-CMD ["uvicorn", "api_server:app", "--host", "0.0.0.0", "--port", "8000", "--reload"] 
+CMD ["uvicorn", "ai_api.api_server:app", "--host", "0.0.0.0", "--port", "8000", "--reload"]
